server/requirements.txt
Django==4.1.13
djongo==1.3.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
pymongo==3.12.3
python-dateutil==2.8.2
pytz==2023.3.post1
sqlparse==0.2.4
urllib3==2.0.7